backend/requirements.txt
# backend/requirements.txt
fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn
pytest==5.4.2
requests==2.23.0
gunicorn==20.0.4
uvloop
httptools
asyncpg==0.20.0
databases[postgresql]==0.2.6
SQLAlchemy==1.3.12
